lecture_notify/backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
bfg==0.1.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.8
Django==5.0.6
django-celery-results==2.5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
filelock==3.15.4
frozenlist==1.4.1
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.7
kombu==5.3.7
multidict==6.0.5
numpy==2.0.0
openpyxl==3.1.4
pandas==2.2.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytoml==0.1.21
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
twilio==9.2.1
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
yarl==1.9.4
dj-database-url>=0.5.0